Add docker conditional back

+++ docker/recipes/docker-deploy.rb
@@ -12,11 +12,13 @@ node[:deploy].each do |application, deploy|
   bash "docker-cleanup" do
     user "root"
     code <<-EOH
+      if docker ps -a 
+      then
         docker stop $(docker ps -a -q)
         sleep 3
         docker rm $(docker ps -a -q)
         sleep 3
-
+      fi
     EOH
   end
 
@@ -26,13 +28,11 @@ node[:deploy].each do |application, deploy|
     user "root"
     #cwd "#{deploy[:deploy_to]}/current"
 
-    #docker run -d -p 54.86.41.97:9292:9292 -p 54.86.41.97:9200:9200 pblittle/docker-logstash
     code <<-EOH
       docker run -d -p 54.86.41.97:9292:9292 -p 54.86.41.97:9200:9200 pblittle/docker-logstash
     EOH
 
     #docker run #{dockerenvs} -p #{node[:opsworks][:instance][:private_ip]}:#{deploy[:environment_variables][:service_port]}:#{deploy[:environment_variables][:container_port]} --name #{deploy[:application]} -d 
-    #docker run -p 10.183.61.222:80:80 -p 10.183.61.222:443:443 -d mydocker
   end
 
 end
